Rename "template" repo to ".github".
This is the default for some apps and is actually required by some other
apps. Very sad.. this also means we can't have .github be a submodule
named .github at the top level of toktok-stack.

As a result, this repo will now sit in tools/.github, so it doesn't
conflict with the actual .github in toktok-stack.