chore: Add size labels for issues.
This will help identify small tasks for people looking to make
contributions but don't want to invest a lot of time, or medium tasks
for people who are willing to spend a bit of time. Large tasks will be a
significant investment and probably need multiple people on it. Extra
large tasks should probably be split up into smaller tasks.